Fix flicker transitioning from LightDM to Openbox
Feh can be replaced with hsetroot now since I just want a solid color.
Feels silly storing a png of solid black for this purpose.

For fix on the screen briefly flickering to grey on login, see:

https://wiki.archlinux.org/title/Openbox

Section Smooth display manager transition

Edited Openbox file isn't tracked in this repo, it's in this path:

/usr/lib/openbox/openbox-autostart