Change PS1 for virtual console (tty2)
It can't display nerdfont icons.

Note: still has an issue with for example showing italics text in Vim,
but by luck the outcome is readable so don't know if I'll bother
checking that in a Vim colorscheme.

tty2 (ctrl+alt+F2) is basically only used for a quick fix when something
went badly wrong with graphical session. Made it a bit nicer to use, but
not too important.