Update firefox menu shadow fix
Only the line `override_redirect = true` was needed afterall.

Noticed that the setting may also disable shadow for some windows that I
want to have it on - in this case it was Rofi.

Two options:

1. Only apply if class_g = 'firefox'
2. Apply if class_g != 'Rofi'

Makes most sense to apply fix only to firefox window, but do the other
if more applications have similar issue.