Hide findbar checkboxes
Prefer to have Highlight All on, the rest off. The options which are
never changed take way too much space on the bar, sometimes can't even
see the count of matches.

Note: can't change the options now that they're hidden, but some (if not
all?) of them are also in about:config (e.g. findbar.highlightAll)