Fix CSD min and max buttons invisible on hover
Gives min, max and close buttons the same bg color (light gray) as the
New Tab button on hover. Setting the color to be the same as when not
hovering was necessary, otherwise min and max will have dark fg color.
Unclear if that's due to my mistake, but I like this solution.