Set LS_COLORS and fix unreadable defaults
Easily editable LS_COLORS in separate file. Some defaults had
foreground-on-foreground colorscheme which is an eye sore. All of
those's foreground set to the bg color (#121212), which corresponds to
`30`.

Since there are .rpmnew and .rpmsave, add .pacnew and .pacsave in the
same color.

This fixes bad colors in fd --color=auto so remove the alias that
disables color in fd