Add keymaps for zathura
Mapping Tab to search for everything that has / to search as default

I would map Shift-Tab to search backward (? search), but sometimes that
specific key combination is difficult to recognize. Is it a "Backtab?"
`xev` shows it as "ISO_Left_Tab", but can't give that to zathura.

Why `focus_inputbar /` instead of `feedkeys /`:

They function the same, but with feedkeys I will get this warning every
time I use the keybind:

    Event with type 8 not holding a GdkDevice. It is most likely
    synthesized outside Gdk/GTK+ # Please enter the commit message for
    your changes.

Since I start zathura from the command line a lot of the time, don't
wanna see this warning.