Re-enable ignorecase and smartcase
Having ignorecase in searches is simply too convenient. The problem
earlier was that it also enables it in :substitute, which is nothing but
a hinderance, for example:

foobar
FooBar
fOObAr

:%s/foobar/bazqux/g
will result in

bazqux
bazqux
bazqux

but what you actually had in mind was

bazqux
BazQux
bAzqUx

so you shot yourself in the foot.

So I suppose in the rare case that this matters, add the I flag in the
end like so:

:%s///Ig

Also take a look at tpope's vim-abolish for further thought on this:

https://github.com/tpope/vim-abolish

The Substitution part is the relevant one.