Rework notifications
While fullscreen, changing volume or brightness shows a fast, stacking
notification because Polybar is not visible.

Detecting fullscreen status may be a bit hacky but it works.

Notifications now on top-right because I could not find a solution to
this:
* When fullscreen, offset from screen edge
* When not fullscreen, offset above polybar

Now what if a notification was displayed when fullscreen, and coming off
fullscreen while it's still visible? Yeah, too complicated.

All of this pain is avoided by using one of the top corners where
there's no bar or nothin'.