Update treesitter highlight groups
Big changes in naming, notably ＠text is now ＠markup and a lot of groups
moved under ＠keyword.<subgroup>, or similarly under ＠variable and so
on.

This allows me to remove a lot of lines because in most cases linking to
the top level (like ＠keyword.repeat -> ＠keyword) is exactly what I
wanted to do.

Full list of changes in the description of this commit:
https://github.com/nvim-treesitter/nvim-treesitter/commit/1ae9b0e4558fe7868f8cda2db65239cfb14836d0