Replace brightnessctl-git with brillo
New commit in brightnessctl makes percentages incredibly broken

The problem with other brightness control programs was always being
unable to do even increments for polybar: 5% would occasionally be 4%.

Printf rounding solved this and I liked `brillo`'s ergonomics.

Note one thing: with `brillo` I can only go as low as 0.39%, and there's
a noticeable difference between 0.00% and 0.39%

But it might not be an issue.. 0.39% works fine for me.