Replace 'pink' with 'burnt sienna'
For the int/bool/const color I went back to orange (#de935f) and worked
a more pleasant variant on it

Using coolors.co I mixed it more with red and "Burnt Sienna" (reddish
brown) is looking really good for ints with the idea that it pops out a
bit more than other colors.

I still name it "orange" in a variable, that's kinda what it's derived
from & which space it occupies in the colorscheme.