Update PS1 and PS2 prompt icons
nf-fa-caret_right  is very similar to unicode 🞂

The fact that these triangles are solid blocks makes them really pop out
especially with the green color. As the nerd icon is slightly smaller,
it feels just a little more proportionate.

I like how it looks when combined to 󱞩 nf-md-arrow-right-bottom as PS2