Update battery module icons
Battery icons are the same but different unicode spot (?)

Changed charging symbol to just lightning bolt (nf-fa-bolt) because the
battery symbols with something small inside  are too hard to distinguish
at such a small size + the moment where discharging turns into charging
and vice versa is now nice to look at because the outline clearly
changes.

Removed `full-at = 99` because battery is reported full at 96 or so, and
the moment it's "full" is clearly seen when the icon is no longer
lightning bolt.