Change rg path color
Default: magenta

Blue makes more sense to me as I think of it kind of like a 'header',
and blue is my color for header in markdown. Also LSCOLOR for
directories. This isn't that great though because the last item of the
path isn't a dir.

Anyways I prefer blue for this.