Change tab behavior on ambiguous case
Can't believe this exists and is not default. By default I'd have to
press tab twice to list all matches, with `show-all-if-unmodified`, I
just have to press tab once. What's the point of the double tab?

More thoughts:

`show-all-if-ambiguous` has benefits, but ultimately is annoying.

Another good choice would be binding Tab to `menu-complete`, to make it
behave like for example Vim's wildmenu.