Add a script to cycle windows of the same class
The code is quite horrible looking but works great + no leaks.

This would have been very easy with `xdo id -c`, but the fact that it
outputs in order of stacking order has some issues; I needed it in order
of _NET_CLIENT_LIST.

Note:
I have some of the `wmctrl` code copypasted in 3 places now, could I
place it somewhere as a library or something like that?