Rename firefox color theme
For some reason, default gnome (?) theme was showing on startup before
loading my Firefox Color theme.

Once again worked around this by loading the theme directly. This time I
had to send it to mozilla as an unlisted addon, can't install the
exported xpi directly with "Install Add-on From File"

So, this works, but if I make changes, changing the version number might
be messy. But I guess it will work out.

So to reiterate: the .xpi on this repo can't be installed directly
before sending to the Addon store (says it's 'corrupted' or whatever).