Improve protonvpn-server module
Found a good example on

https://github.com/polybar/polybar/wiki/Module:-script#examples

with the `exec-if` which simplifies the script: no need to start
grepping and cutting `nmcli` outputs if `openvpn` is not running.

Add click action on the module

Not usable when VPN isn't on, but that's by design: in that case use
rofi to start protonvpn-app.