alacritty: make font size much smaller
To see more lines at once, avoid too much need for scrolling, better
default.

Not quite sure how the size works (especially decimals), but the idea
was to reduce it to the next point where a full-width window doesn't
have a large gap between the last terminal cell and polybar.

One annoyance here is that zooming in with Ctrl + + never quite hits the
previous value of "size 12.5" (which would be a nice bigger size
situationally, that avoids the bottom gap)

Question: what exactly do Ctrl + + and Ctrl + - do (what's the value of
the inc/decrement)?