nvim: add mappings for `gg` and `G`
After making the custom scroll function (38f2990), I noticed I had a
habit with `<C-d>` and `<C-u>` to spam it to get the cursor to
last/first line of buffer.

The point of the new scroll method is to prevent it accidentally going
there.

Solution: use `G` and `gg` instead for that purpose.

But they're annoyingly frictionful, e.g. try spamming
    `ggGggGggGggG`
    vs.
    `<M-BS><M-Enter><M-BS><M-Enter><M-BS><M-Enter><M-BS><M-Enter>`

Especially makes nvimpager nicer to use.