nvim: rework `search-improvements.lua`
It was getting unbearably hacky.

While this was helpful:

https://www.reddit.com/r/neovim/comments/zc720y/tip_to_manage_hlsearch/

it confused some things such as:
    - when `vim.o.hlsearch` is on, the builtin search commands (/?nN*#)
      turn `v:hlsearch` on (note: not the option!)
    - to turn off, either `:nohlsearch` or `:let v:hlsearch = 0`

So we never need to handle turning the hlsearch *on*.

For caveats on turning hlsearch off in *autocommands*:
    - `:h :nohlsearch` `autocmd-searchpat`
    - `:h v:hlsearch`, `:h function-search-undo`

^ Looks like I managed to dodge these because the `:nohlsearch` is used
  in the `on_key` listener (technically outside the autocmd)?
    - (I'm not sure of this)

Get rid of the auto-centering in favor off `set scrolloff = 6`. Re-add
if that feels too low for search cycling.