nvim: revert to builtin `x` keymap
Was experimenting:

Since I consider builtin `s` remappable, shouldn't the same apply to
`x`?

Well, the conclusion is hell no: I use `x` very, very frequently. It's
so common to for example delete a word, but still need that 1 character
removed because it doesn't quite cut it, so do a single `x` afterwards.

For this purpose, `<M-d>` felt considerably more annoying to use.

Return to tommcdo-like exchange-prefix.