nvim: remove 'anybracket' and 'anyquote' textobjs
Why:

Want to always be exact and not have to consider different brackets as
potential obstacles, at all.

That said, I found 'anyquote' frequently useful, until I set up the
exact 'aliases' for each (of the three) quotes, I find it no longer
necessary.

With 'anybracket' it was common to trip up with nested {} and ()
particularly. I think the aliases are a better idea.

Later, I'll refine the insert mode mappings, which use these quote- and
bracket-aliases, so it's extra important to always be typing the exact
alias.