nvim: remove `tildeop`
Turns out the way I really ever used a switch case operator was `äl` so
just make it a single keypress again.

I was interested in making it not move to the right with something like:

    nnoremap ä v~

but it would require further thought for using with a count, and a
separate visual mode mapping.