firefox/nvim: add Textern addon
Addon to edit browser text fields in nvim:
  https://github.com/jlebon/textern
  https://addons.mozilla.org/en-US/firefox/addon/textern/

Requires:
  https://aur.archlinux.org/packages/firefox-extension-textern-native-git

The titlestring when using textern was like a million characters long,
the main thing we wanna see in the title when using Textern, is that it
was launched by Textern: it's a special kind of nvim session. Buffername
we don't wanna see.

Tiny nitpick: when buffer isn't 'modified', there shouldn't be a space
after [Textern]