polybar: add USB-C headphones plug detection (hacky)
First tried by making udev rules, but could not find a way to update the
icon and volume value correctly when they are plugged -- and there's a
small delay until the sound actually comes from the headphones. However,
even adding a `sleep 10` before running the script didn't work in this
context, which is so weird and frustrating.

Here's a hacky alternative that tails `pactl subscribe`, which to be
fair I've seen many other custom Polybar scripts do.

All in all feels a little bit heavy-weight but still worth it over the
builtin `pulseaudio` module.

Should be improved at the first sign of any ideas.