zsh/nvim: separate bright black and comment color
The comment color is too dark/concealed in many cases where it was used
as ANSI "bright black" by some shell programs. Better have a dedicated
neovim comment color and a brighter bright black.

This leaves a duplicate:

"darker darker fg" and "bright black" are the same

but as the "darker darker fg" is more descriptive and useful in the
neovim color scheme, leaving it in for now.