Revert "nvim: fix md codeblock textobject"
Revert 9d9d224

The solution was not complete, mostly because I want the selection to
always be linewise, even if the md codeblock only has one line in it
(which is common).

Todo: Considering rewriting this as a full function the same way
`ai_indent` was made (if a more concise solution doesn't exist).