nvim: reimplement mini.ai md codeblock textobject
This needed full control mainly for this reason: I wanted the `i`-type
object to always operate linewise, even if the codeblock has only one
line (which is common).

For `a` it's more complicated to be linewise, for example because md
codeblocks can be in bulleted lists:

- ```
  foo
  ```

Would not be proper to include the bullet in this plus there are more
similar cases.

Made it so `a` is linewise if the codeblock sits on col 1.