firefox: quick fix for messed up window control buttons
Firefox update caused:
1. enormous window control buttons
2. unwanted background color on them

With the new css vars like `--titlebar-button-size` it was actually
easier to try to get it to match my Openbox theme almost
pixel-perfectly. But the biggest issue is: the current `restore.svg`
doesn't fit a size this small: it gets a little bit malformed.

Todo: replace `restore.svg` with a new icon that's drawn more like
      `max_toggled.xbm` from my Openbox theme.