nvim/zsh: add another map to get out of pairs
It's always felt so bad to do for example:

    Alt + d, type some text, Alt + l

And this happens so often, I needed a second map for the Alt + l that
'flows' better in that context

Alt + space feels promising, especially inserting and escaping empty
pairs:

    Hold alt, d, Space

This sequence feels similar to using my Alt + Enter map (which starts a
new line inside the pairs)

Unfortunately had to replace previous Alt + Space which made sense and
was useful, as "surround cursor with spaces"

But I've only ever very rarely used that. Bind to something else, if it
feels needed.

Considered Shift + Alt + Space for backward variant, but in zsh that key
combo cannot be used. In nvim it can, but didn't add it, to be consistent.