openbox: change volume/brightness keybinds
Turns out it's important that these can be done one-handedly.

Grouped all of these on F1-F4 regardless of what the label on the key is.

Now I can just place my hand on the first F-key cluster and use the
first 2 / last 2, instead of having to feel for the middle 2, which was
the case when matching the label for volume up/down (F2, F3).

Removed chords with 2 modifiers and used the ones that are easiest to
combo with F1-F2.