nvim: workaround custom <Esc> map and sneak#cancel
Not the most consistent solution, for example:

1. in Sneak, pressing <Esc> both cancels sneak *and* clears cmdline
2. in mini.operators.exchange, <Esc> *only* cancels exchange, until the
   next press

Would prefer all to be like 2, but it doesn't matter that much.

Todo: maybe hack on Sneak so it can handle this, or improve the keymap.
I already tried an if-clause with "if sneak#is_sneaking"... but it
didn't work how I expected in label mode (would take another press of
<Esc> until I reach my own custom keybind).