nvim: finally disable matchparen builtin plugin
There was a problem with Sneak where disabling Matchparen caused the
Sneak highlights to clear later than intended (seemed to be related to
autocmd CursorMoved, but it could be a false flag).

I've reworked Sneak highlight behavior such that they don't clear on
autocmds anymore.

Plus, would be better to take a proper look at why Sneak has this odd
dependence on Matchparen.