nvim: try `0` mapping on `<leader>h`
<M-Esc> is a little bit too acrobatic oftentimes (especially
operator-pending), although became muscle memory fairly well.

Since <leader>h is free, why not that?

I suppose it would be nice for <leader>h and <leader>l to be something
symmetric,

But that's not a problem I have today.

The clever insert-mode-escape-variant was cool on paper but there was
never a situation where I would want to use that.